+++ score/weather.go
@@ -2,7 +2,7 @@ package score
 
 import "github.com/nkzren/ecoscheduler/weather"
 
-func weatherScore(loc weather.Location) (int, error) {
+func weatherScore(loc weather.Location) (float64, error) {
 	data, err := weather.GetData(loc)
 	if err != nil {
 		return -1, err
@@ -10,6 +10,6 @@ func weatherScore(loc weather.Location) (int, error) {
 	return calculateScore(data), nil
 }
 
-func calculateScore(data *weather.WeatherbitData) int {
-	return int(weather.GetWindEnergy(data) + weather.GetSunEnergy(data))
+func calculateScore(data *weather.WeatherbitData) float64 {
+	return weather.GetWindEnergy(data) + weather.GetSunEnergy(data)
 }
